Use `ll` in IterativeLazySegmentTree

+++ stress-tests/data-structures/IterativeLazySegmentTree.cpp
@@ -4,8 +4,8 @@
 
 namespace Slow {
 struct Tree {
-	typedef int T; // data type
-	struct L { int mset, madd; }; // lazy type
+	typedef ll T; // data type
+	struct L { ll mset, madd; }; // lazy type
 	const T tneut = -inf;		 // neutral elements
 	const L lneut = {inf, 0};
 	T f (T a, T b) { return max(a, b); } // (any associative fn)
@@ -17,17 +17,17 @@ struct Tree {
 		return {a.mset, a.madd + b.madd};
 	} // Combine lazy
 
-	int n;
+	ll n;
 	vector<T> vals;
 
   Tree(const vector<T> vals) : n(sz(vals)), vals(vals) {}
 
-  void update(int b, int e, L a) { // update [b, e)
+  void update(ll b, ll e, L a) { // update [b, e)
     rep(i, b, e) {
       vals[i] = apply(vals[i], a);
     }
   }
-  T query(int b, int e) { // query [b, e)
+  T query(ll b, ll e) { // query [b, e)
     T ans = tneut;
     rep(i, b, e) {
       ans = f(ans, vals[i]);
@@ -44,9 +44,9 @@ int main() {
   }
 
   rep(_, 0, 1000) {
-    int n = rand() % 100 + 1;
+    ll n = rand() % 100 + 1;
 
-    vector<int> vals(n);
+    vector<ll> vals(n);
     rep(i, 0, n) {
       vals[i] = rand() % 100;
     }
@@ -57,7 +57,7 @@ int main() {
     vi as, bs, tys;
 
     rep(__, 0, 100) {
-      int a = rand() % n, b = rand() % n, ty = rand() % 3;
+      ll a = rand() % n, b = rand() % n, ty = rand() % 3;
 
       if (b == a) b++;
       if (b < a) swap(a, b);
@@ -65,7 +65,7 @@ int main() {
       as.push_back(a), bs.push_back(b), tys.push_back(ty);
 
       if (ty <= 1) {
-        int x = rand() % 100;
+        ll x = rand() % 100;
 
         if (ty == 0) {
           t.update(a, b, {x, 0});
@@ -75,12 +75,12 @@ int main() {
           t2.update(a, b, {inf, x});
         }
       } else {
-        int ans = t.query(a, b);
-        int ans2 = t2.query(a, b);
+        ll ans = t.query(a, b);
+        ll ans2 = t2.query(a, b);
         assert(ans == ans2);
       }
     }
-    int a = rand() % n, b = rand() % n;
+    ll a = rand() % n, b = rand() % n;
     if (b == a) b++;
     if (b < a) swap(a, b);
   }
